Crate slosh_test_lib

Modules

docs

Functions

add_user_builtins

do_to_vm_stdout_fcns

pr/prn/dasm/dump-globals which write directly to stdout are handed new behavior according to variable: NoopSwap.

vm_with_builtins_and_core

If noop_stdout is set to true then all functions that write to stdout (pr/prn/dasm/dump-globals) will be overwritten with the noop function.

vm_with_stdout_disabled

pr/prn/dasm/dump-globals which write directly to stdout are mapped to noop

vm_with_stdout_enabled

pr/prn/dasm/dump-globals which write directly to stdout are not mapped to noop