slosh_test_lib::docs

Function add_user_docs_to_mdbook_less_provided_sections

pub fn add_user_docs_to_mdbook_less_provided_sections(
    vm: &mut SloshVm,
    md_book: &mut Book,
    provided_sections: BTreeMap<String, BTreeSet<SloshDoc>>,
) -> VMResult<()>

allows user specific forms to be identified. everything gathered recently. not in provided sections

provided sections will have a lot of information that the return of get_docs_by_section will have but the new invocation should have more, MUST call a version of the vm that includes init.slosh

this method is very specific and should be provided a list of sections gathered without init.slosh loaded and then called after its loaded.