slosh_test_lib::docs::legacy

Function full_legacy_sl_sh_forms_metadata

pub(crate) fn full_legacy_sl_sh_forms_metadata() -> &'static [(&'static str, bool, &'static str)]

Array of triples: 1. String name of the sl-sh form 2. Whether or not it will be implemented in slosh 3. If 2 is true: an optional non-null value listing the name the function is mapped to in slosh if the name has changed. If 2 is false: a note a helpful explanation as to why the function is no longer needed or will not be implemented. ** all values in .0 (or values in .2 if .1 is true) are assumed to match to some member of the set returned by (get-globals-sorted), any values not found in this set are assumed to be unimplemented.