slosh_test_lib
Module docs

Modules

legacy

Structs

DOC_REGEX ğŸ”’

DocStringSection ğŸ”’

EXEMPTIONS

SloshDoc ğŸ”’

Enums

DocError ğŸ”’

Namespace ğŸ”’

Constants

DESCRIPTION ğŸ”’

EXAMPLE ğŸ”’

SECTION ğŸ”’

USAGE ğŸ”’

Traits

AsMd

Functions

add_builtins

build_all_slosh_forms_listing_chapter ğŸ”’

build_doc ğŸ”’

build_each_docs_section_chapter ğŸ”’

build_sl_sh_transition_chapter ğŸ”’

build_symbols_list ğŸ”’

capitalize_first ğŸ”’

doc_map ğŸ”’

get_docs_by_section ğŸ”’

Each doc has a tag in its Section: definition by convention that logically groups functions. Using a HashMap store the section tags as keys and add all slosh docs from to a vector as a value corresponding to its section.

get_exemptions ğŸ”’

get_globals_sorted ğŸ”’

insert_nil_section ğŸ”’

insert_section ğŸ”’

make_file ğŸ”’

name_for_all_page ğŸ”’

name_for_section_page ğŸ”’

symbol_and_capitalized_symbol ğŸ”’

Type Aliases

DocResult ğŸ”’