slosh_test_lib

Module docs

Modules

legacy

Structs

EXEMPTIONS

SloshDoc

Traits

AsMd

Functions

add_builtins

add_forms_and_supplementary_docs

Convention is that the section that enumerates all of the forms by section (get_slosh_docs) is added to MDBook and then the supplementary materials (link_supplementary_docs).

add_user_docs_to_mdbook_less_provided_sections

allows user specific forms to be identified. everything gathered recently. not in provided sections

get_slosh_docs

Retrieve the docs for each section. Optional side-effects write docs to provided Book if toggled with bool).

link_supplementary_docs

Supplementary materials include things like, the sl-sh -> slosh port tracker, actual Rust documentation, and legacy sl-sh documentation.