slosh_test_lib

Module docs

Modules

legacy

Structs

DOC_REGEX ğŸ”’

DocSearchQuery ğŸ”’

DocStringSection ğŸ”’

EXEMPTIONS

SloshDoc

StyleOptions ğŸ”’

Enums

DocError ğŸ”’

Namespace ğŸ”’

Constants

DESCRIPTION ğŸ”’

EXAMPLE ğŸ”’

GLOBAL_NAMESPACE ğŸ”’

SECTION ğŸ”’

USAGE ğŸ”’

USER_FORMS ğŸ”’

Traits

AsMd

Functions

add_builtins

add_forms_and_supplementary_docs

Convention is that the section that enumerates all of the forms by section (get_slosh_docs) is added to [MDBook] and then the supplementary materials (link_supplementary_docs).

add_forms_to_md_book_part ğŸ”’

add_user_docs_to_mdbook_less_provided_sections

allows user specific forms to be identified. everything gathered recently. not in provided sections

basic_match ğŸ”’

build_all_slosh_forms_listing_chapter ğŸ”’

build_doc ğŸ”’

load md book and attach doc to it. can also attach pre-processors.

build_each_docs_section_chapter ğŸ”’

build_sl_sh_transition_chapter ğŸ”’

build_symbols_list ğŸ”’

capitalize_first ğŸ”’

doc_map ğŸ”’

doc_search ğŸ”’

generate_styled_output ğŸ”’

get_docs_by_section ğŸ”’

Each doc has a tag in its Section: definition by convention that logically groups functions. Using a HashMap store the section tags as keys and add all slosh docs from to a vector as a value corresponding to its section.

get_exemptions ğŸ”’

get_globals_sorted ğŸ”’

get_slosh_docs

Retrieve the docs for each section. Optional side-effects write docs to provided [Book] if toggled with bool).

highlight_basic_matches ğŸ”’

highlight_matches ğŸ”’

highlight_regex_matches ğŸ”’

highlight_substring_matches ğŸ”’

insert_nil_section ğŸ”’

insert_section ğŸ”’

legacy_report ğŸ”’

link_supplementary_docs

Supplementary materials include things like, the sl-sh -> slosh port tracker, actual Rust documentation, and legacy sl-sh documentation.

make_file ğŸ”’

name_for_all_page ğŸ”’

name_for_section_page ğŸ”’

search_docs ğŸ”’

symbol_and_capitalized_symbol ğŸ”’

Type Aliases

DocResult ğŸ”’